refactor(src): restructure state packages
